+++ slackclient/slackrequest.py
@@ -1,8 +1,6 @@
-import json
-
 import requests
-import six  # noqa: F401
-
+import json
+import six
 import sys
 import platform
 from .version import __version__
@@ -62,13 +60,13 @@ class SlackRequest(object):
 
         url = 'https://{0}/api/{1}'.format(domain, request)
 
-        # Override token header if token is passed in form params
+        # Override token header if `token` is passed in post_data
         if post_data is not None and "token" in post_data:
             token = post_data['token']
 
+        # Set user-agent and auth headers
         headers = {
             'user-agent': self.get_user_agent(),
-            'Content-Type': 'application/json',
             'Authorization': 'Bearer {}'.format(token)
         }
 
@@ -77,22 +75,30 @@ class SlackRequest(object):
         # use the 'file' argument to point to a File ID.
         post_data = post_data or {}
 
+        # Move singular file objects into `files`
+        upload_requests = ['files.upload']
+
+        # Move file content into requests' `files` param
+        files = None
+        if request in upload_requests:
+            files = {'file': post_data.pop('file')} if 'file' in post_data else None
+
         # Check for plural fields and convert them to comma-separated strings if needed
         for field in {'channels', 'users', 'types'} & set(post_data.keys()):
             if isinstance(post_data[field], list):
                 post_data[field] = ",".join(post_data[field])
 
-        # Move singular file objects into `files`
-        upload_requests = ['files.upload']
+        # Convert any params which are list-like to JSON strings
+        # Example: `attachments` is a dict, and needs to be passed as JSON
+        for k, v in six.iteritems(post_data):
+            if isinstance(v, (list, dict)):
+                post_data[k] = json.dumps(v)
 
-        files = None
-        if request in upload_requests:
-            files = {'file': post_data.pop('file')} if 'file' in post_data else None
-
+        # Submit the request
         return requests.post(
             url,
             headers=headers,
-            data=json.dumps(post_data),
+            data=post_data,
             files=files,
             timeout=timeout,
             proxies=self.proxies