+++ tests/test_slackrequest.py
@@ -41,6 +41,7 @@ def test_auth_header(mocker):
 
     assert "Bearer xoxb-123" in kwargs['headers']['Authorization']
 
+
 def test_token_override(mocker):
     requests = mocker.patch('slackclient.slackrequest.requests')
     request = SlackRequest()
@@ -63,9 +64,12 @@ def test_plural_field(mocker):
     request.do('xoxb-123','conversations.open', {'users': ['U123', 'U234', 'U345']})
     args, kwargs = requests.post.call_args
 
-    assert requests.post.call_count == 1
-    assert 'https://slack.com/api/conversations.open' == args[0]
-    assert kwargs['data'] == '{"users": "U123,U234,U345"}'
+    assert kwargs['data'] == {'users': 'U123,U234,U345'}
+
+    request.do('xoxb-123','conversations.open', {'users': "U123,U234,U345"})
+    args2, kwargs2 = requests.post.call_args
+
+    assert kwargs2['data'] == {'users': 'U123,U234,U345'}
 
 
 def test_post_file(mocker):
@@ -80,7 +84,7 @@ def test_post_file(mocker):
 
     assert requests.post.call_count == 1
     assert 'https://slack.com/api/files.upload' == args[0]
-    assert {'filename': 'slack_logo.png'} == json.loads(kwargs['data'])
+    assert {'filename': 'slack_logo.png'} == kwargs['data']
     assert kwargs['files'] is not None
 
 
@@ -93,7 +97,7 @@ def test_get_file(mocker):
 
     assert requests.post.call_count == 1
     assert 'https://slack.com/api/files.info' == args[0]
-    assert {'file': "myFavoriteFileID"} == json.loads(kwargs['data'])
+    assert {'file': "myFavoriteFileID"} == kwargs['data']
     assert kwargs['files'] is None
 
 
@@ -108,5 +112,4 @@ def test_post_attachements(mocker):
 
     assert requests.post.call_count == 1
     assert 'https://slack.com/api/chat.postMessage' == args[0]
-    assert {'attachments': [{'title': 'hello'}]} == json.loads(kwargs['data'])
-    assert kwargs['files'] is None
+    assert isinstance(kwargs["data"]["attachments"], str)